+++ src/main/java/insane96mcp/iguanatweaksreborn/IguanaTweaksReborn.java
@@ -11,7 +11,6 @@ import insane96mcp.iguanatweaksreborn.data.generator.client.ITRBlockModelsProvid
 import insane96mcp.iguanatweaksreborn.data.generator.client.ITRBlockStatesProvider;
 import insane96mcp.iguanatweaksreborn.data.generator.client.ITRItemModelsProvider;
 import insane96mcp.iguanatweaksreborn.modifier.Modifiers;
-import insane96mcp.iguanatweaksreborn.module.combat.InnateResistance;
 import insane96mcp.iguanatweaksreborn.module.combat.PiercingPickaxes;
 import insane96mcp.iguanatweaksreborn.module.combat.RegeneratingAbsorption;
 import insane96mcp.iguanatweaksreborn.module.experience.anvils.AnvilRepairReloadListener;
@@ -89,7 +88,6 @@ public class IguanaTweaksReborn
         modEventBus.addListener(this::addPackFinders);
         modEventBus.addListener(PiercingPickaxes::addAttribute);
         modEventBus.addListener(RegeneratingAbsorption::addAttribute);
-        modEventBus.addListener(InnateResistance::addAttribute);
         modEventBus.register(Stamina.class);
 		modEventBus.register(SpawnerData.class);
         modEventBus.register(Tiredness.class);